[fix] 完善

+++ src/pyosis/load/manager.py
@@ -336,7 +336,7 @@ class LoadCase:
             bY: int = 0,
             Dy: float = 0.0,
             bZ: int = 0,
-            Dz: int = 0,
+            Dz: float = 0.0,
             bRx: int = 0,
             Rx: float = 0.0,
             bRy: int = 0,
